[WIP] Platform Traling Stop

+++ database/migrations/2022_01_05_183000_wallet_buy_sell_exchange.php
@@ -8,13 +8,14 @@ return new class extends MigrationAbstract {
     /**
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         if ($this->upMigrated()) {
             return;
         }
 
-        $this->tables();
+        $this->upTables();
+        $this->upUpdate();
     }
 
     /**
@@ -29,35 +30,48 @@ return new class extends MigrationAbstract {
     /**
      * @return void
      */
-    protected function tables()
+    protected function upTables(): void
     {
         Schema::table('wallet', function (Blueprint $table) {
-            $table->unsignedDouble('sell_stop_exchange')->default(0)->after('sell_stop');
-            $table->unsignedDouble('buy_stop_exchange')->default(0)->after('buy_stop');
+            $table->unsignedDouble('sell_stop_exchange')->default(0);
+            $table->unsignedDouble('buy_stop_exchange')->default(0);
         });
 
-        Schema::table('wallet_history', function (Blueprint $table) {
-            $table->unsignedDouble('sell_stop_exchange')->default(0)->after('sell_stop');
-            $table->unsignedDouble('buy_stop_exchange')->default(0)->after('buy_stop');
-        });
-
-        $this->db()->statement('UPDATE `wallet` SET `sell_stop_exchange` = `buy_exchange`, `buy_stop_exchange` = `buy_exchange`;');
-        $this->db()->statement('UPDATE `wallet_history` SET `sell_stop_exchange` = `buy_exchange`, `buy_stop_exchange` = `buy_exchange`;');
+        if (Schema::hasColumn('wallet_history', 'buy_exchange')) {
+            Schema::table('wallet_history', function (Blueprint $table) {
+                $table->unsignedDouble('sell_stop_exchange')->default(0);
+                $table->unsignedDouble('buy_stop_exchange')->default(0);
+            });
+        }
     }
 
     /**
      * @return void
      */
-    public function down()
+    protected function upUpdate(): void
+    {
+        $this->db()->statement('UPDATE `wallet` SET `sell_stop_exchange` = `buy_exchange`, `buy_stop_exchange` = `buy_exchange`;');
+
+        if (Schema::hasColumn('wallet_history', 'buy_exchange')) {
+            $this->db()->statement('UPDATE `wallet_history` SET `sell_stop_exchange` = `buy_exchange`, `buy_stop_exchange` = `buy_exchange`;');
+        }
+    }
+
+    /**
+     * @return void
+     */
+    public function down(): void
     {
         Schema::table('wallet', function (Blueprint $table) {
             $table->dropColumn('sell_stop_exchange');
             $table->dropColumn('buy_stop_exchange');
         });
 
-        Schema::table('wallet_history', function (Blueprint $table) {
-            $table->dropColumn('sell_stop_exchange');
-            $table->dropColumn('buy_stop_exchange');
-        });
+        if (Schema::hasColumn('wallet_history', 'sell_stop_exchange')) {
+            Schema::table('wallet_history', function (Blueprint $table) {
+                $table->dropColumn('sell_stop_exchange');
+                $table->dropColumn('buy_stop_exchange');
+            });
+        }
     }
 };
